kvs: pull benchmark into separate file (#116)
Also remove the reporting of latency information for now, since it was
vulnerable to coordinated omission and is not reliable.
